vividus-engine/src/main/java/org/vividus/replacement/DeprecatedCodeStepsReporter.java
/*
 * Copyright 2019-2023 the original author or authors.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     https://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package org.vividus.replacement;

import java.lang.reflect.Method;
import java.util.Optional;

import org.jbehave.core.steps.NullStepMonitor;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.vividus.annotation.Replacement;

public class DeprecatedCodeStepsReporter extends NullStepMonitor
{
    private static final Logger LOGGER = LoggerFactory.getLogger(DeprecatedCodeStepsReporter.class);

    private final DeprecatedStepNotificationFactory deprecatedStepNotificationFactory;

    public DeprecatedCodeStepsReporter(DeprecatedStepNotificationFactory deprecatedStepNotificationFactory)
    {
        this.deprecatedStepNotificationFactory = deprecatedStepNotificationFactory;
    }

    @Override
    public void beforePerforming(String step, boolean dryRun, Method method)
    {
        Optional.ofNullable(method).map(m -> method.getAnnotation(Replacement.class))
                .ifPresent(r -> LOGGER.info(deprecatedStepNotificationFactory
                        .createDeprecatedStepNotification(r.versionToRemoveStep(), r.replacementFormatPattern())));
    }
}
